+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Notable changes to this project will be documented in this file
 
+- [X] Prevent users from setting or updating their own roles - 7/13/22
+- [X] Add pagination to the users get all endpoint - 7/13/22
+- [X] Add a update user role endpoint & add guards - 7/13/22
+- [X] Add guards to get all users endpoint - 7/13/22
+- [X] Initial implementation of user roles and guards - 7/13/22
 - [X] Fixed login endpoint, update postman collection, and clarify documentation - 7/7/22
 - [X] Updated Documentation & Postman collection - 7/6/22
 - [X] Fixed PUT update endpoint not working due to params being wrong - 7/6/22